chore: move `promptForAddition` & avoid side effect
