+++ src/commonMain/kotlin/teksturepako/pakku/cli/arg/Prompt.kt
@@ -8,10 +8,15 @@ import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
 import com.github.michaelbull.result.getOrElse
 import teksturepako.pakku.api.actions.ActionError
+import teksturepako.pakku.api.actions.RequestHandlers
 import teksturepako.pakku.api.data.LockFile
 import teksturepako.pakku.api.platforms.GitHub
+import teksturepako.pakku.api.platforms.Platform
 import teksturepako.pakku.api.platforms.Provider
 import teksturepako.pakku.api.projects.Project
+import teksturepako.pakku.cli.resolveDependencies
+import teksturepako.pakku.cli.ui.getFullMsg
+import teksturepako.pakku.cli.ui.pSuccess
 
 suspend fun promptForProject(
     provider: Provider, terminal: Terminal, lockFile: LockFile, fileId: String? = null
@@ -42,4 +47,32 @@ var overrideYes = false
 fun ynPrompt(prompt: String, terminal: Terminal, default: Boolean? = null): Boolean
 {
     return if (overrideYes) true else YesNoPrompt(prompt, terminal, default).ask() == true
+}
+
+suspend fun Project.promptForAddition(
+    lockFile: LockFile,
+    terminal: Terminal,
+    isRecommended: Boolean,
+    noDepsFlag: Boolean,
+    reqHandlers: RequestHandlers,
+    projectProvider: Provider,
+    platforms: List<Platform>
+)
+{
+    val projMsg = this.getFullMsg()
+    val oldProject = lockFile.getProject(this)
+    val replacing = oldProject != null
+    if (ynPrompt("Do you want to ${if (replacing) "replace" else "add"} $projMsg?", terminal, isRecommended))
+    {
+        if (replacing) lockFile.update(this) else lockFile.add(this)
+
+        lockFile.linkProjectToDependents(this)
+
+        if (!noDepsFlag)
+        {
+            this.resolveDependencies(terminal, reqHandlers, lockFile, projectProvider, platforms)
+        }
+
+        terminal.pSuccess("$projMsg ${if (replacing) "replaced" else "added"}")
+    }
 }
